Further decoupled tweening update code
Tweening.update() took a value for the entity as what to do after the tween is complete. Now whatever object is being tweened has a property called tweening, and when it's tweening it sets that value to 1, when it's not it sets it to 0. I feel maybe it'd be better to use a delete or something just for the property itself, but maybe not.